correct currency bugs
