+++ src/component/CourseTrainingCard/index.jsx
@@ -1,6 +1,8 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import styles from "./styles.module.css";
+import { useEffect } from "react";
+import axios from "axios";
 
 const CourseTrainingCard = ({
   id,
@@ -8,11 +10,48 @@ const CourseTrainingCard = ({
   title,
   category,
   price,
-  currency,
   level,
   rating,
   type,
 }) => {
+  const [currency, setCurrency] = useState(null);
+  const [error, setError] = useState(null);
+  const currencies = {
+  "Algeria": { currency: "Algerian dinar", code: "DZD" },
+  "Belgium": { currency: "Euro", code: "EUR" },
+  "Canada": { currency: "Canadian dollar", code: "CAD" },
+  "France": { currency: "Euro", code: "EUR" },
+  "Germany": { currency: "Euro", code: "EUR" },
+  "Morocco": { currency: "Moroccan dirham", code: "MAD" }, 
+  "Tunisia": { currency: "Tunisian dinar", code: "TND" },
+  "Egypt": { currency: "Egyptian pound", code: "EGP" },
+  "United Kingdom": { currency: "Pound sterling", code: "GBP" },
+  "United States": { currency: "United States dollar", code: "USD" },
+  };
+
+  useEffect(() => {
+    const fetchCurrency = async () => {
+      try {
+      
+        const ipResponse = await axios.get('https://api.ipify.org?format=json');
+        const ip = ipResponse.data.ip; 
+        const countryResponse = await axios.get(`https://ipapi.co/${ip}/json/`);
+        const country = countryResponse.data.country_name;
+        const currencyData = currencies[country];
+
+        if (currencyData) {
+          setCurrency(currencyData);
+        } else {
+          setError('Monnaie non trouvée pour ce pays');
+        }
+      } catch (err) {
+        setError('Erreur lors de la récupération des données');
+        console.error(err);
+      }
+    };
+
+    fetchCurrency();
+  }, []);
   return (
     <Link to={`/${type}/${id}`} key={id}>
       <div key={id} className={styles.container}>
@@ -25,7 +64,7 @@ const CourseTrainingCard = ({
         <div className={styles.content}>
           <div className={styles.text}>
             <p>{category}</p>
-            <p>{price} {currency}</p>
+            <p>{price} {currency?.code}</p>
           </div>
 
           <div className={styles.title}>
@@ -33,7 +72,7 @@ const CourseTrainingCard = ({
             <p>{title}</p>
           </div>
 
-          <div className={styles.rating}>
+          {/* <div className={styles.rating}>
             <span>⭐ {rating} (750)</span>
             <div className={styles.stars}>
               <div className={styles.avatarGroup}>
@@ -52,7 +91,7 @@ const CourseTrainingCard = ({
               </div>
               <span>3k+</span>
             </div>
-          </div>
+          </div> */}
 
           <div className={styles.type}>
             <img src="./images/home/type.png" alt="" />
