+++ src/component/ProfileClient/components/CourseCardElement/index.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useRef } from "react";
+import React, { useState,useEffect, useRef } from "react";
 import styles from "./styles.module.css";
 import useCategories from "../../hooks/use-categories";
 import useProfile from "../../hooks/use-profile";
@@ -34,7 +34,44 @@ export default function CourseElement({ course, type }) {
     setOpenPopup(!openPopup);
   };
   const navigate = useNavigate();
+  const [currency, setCurrency] = useState(null);
+  const [error, setError] = useState(null);
+  const currencies = {
+  "Algeria": { currency: "Algerian dinar", code: "DZD" },
+  "Belgium": { currency: "Euro", code: "EUR" },
+  "Canada": { currency: "Canadian dollar", code: "CAD" },
+  "France": { currency: "Euro", code: "EUR" },
+  "Germany": { currency: "Euro", code: "EUR" },
+  "Morocco": { currency: "Moroccan dirham", code: "MAD" }, 
+  "Tunisia": { currency: "Tunisian dinar", code: "TND" },
+  "Egypt": { currency: "Egyptian pound", code: "EGP" },
+  "United Kingdom": { currency: "Pound sterling", code: "GBP" },
+  "United States": { currency: "United States dollar", code: "USD" },
+  };
 
+  useEffect(() => {
+    const fetchCurrency = async () => {
+      try {
+      
+        const ipResponse = await axios.get('https://api.ipify.org?format=json');
+        const ip = ipResponse.data.ip; 
+        const countryResponse = await axios.get(`https://ipapi.co/${ip}/json/`);
+        const country = countryResponse.data.country_name;
+        const currencyData = currencies[country];
+
+        if (currencyData) {
+          setCurrency(currencyData);
+        } else {
+          setError('Monnaie non trouvée pour ce pays');
+        }
+      } catch (err) {
+        setError('Erreur lors de la récupération des données');
+        console.error(err);
+      }
+    };
+
+    fetchCurrency();
+  }, []);
   return (
     <div className={styles.courseContainerElement}>
       <div className={styles.imgCourseContainer}>
@@ -90,7 +127,7 @@ export default function CourseElement({ course, type }) {
  
           > */}
           <div className={styles.buttonsContainer}>
-            <p className={styles.price}>{course?.Price}DT</p>
+            <p className={styles.price}>{course?.Price}{currency?.code}</p>
             <div className={styles.underline} />
             {data?.CoursesPaid?.includes(course._id) ? (
               <div className={styles.statePrimary}>
